Apple client: extract shared VTDecoderBase for H.264/H.265
H264Decoder was ~70% a copy of H265Decoder. Move the shared VideoToolbox
plumbing — NSLock-guarded state, start-code splitting, decompression-session
creation, AVCC packing, and the async decode callback — into VTDecoderBase.
Each decoder keeps only its codec-specific parameter-set layout (SPS/PPS vs
VPS/SPS/PPS) and format-description constructor.

Behavior-preserving: H.265 still decodes each VCL NAL as its own access unit
(decodeAccessUnit([nal])), producing byte-identical AVCC and the same
per-slice decode calls as before; H.264 keeps its multi-slice batching.
